Rename OptimizerRuleReplaceAnyEqWithIn to ReplaceAnyEqWithIn to match naming convention
